+++ source/PubSubHubbub/Models/Namespaces.cs
@@ -1,0 +1,8 @@
+﻿namespace PubSubHubbub.Models;
+
+public static class Namespaces
+{
+    public const string Atom = "http://www.w3.org/2005/Atom";
+    public const string YouTube = "http://www.youtube.com/xml/schemas/2015";
+    public const string Tombstones = "http://purl.org/atompub/tombstones/1.0";
+}